Added audio clips, implemented audio, added new tags.
I added new audio clips for the future and for what i was currently working on.
implemented audio for clicking on clue interactables.
Implemented audio for breaking objects.
Implemented new audio for the prison cell door and debris.
Implemented audio for the loose objects and eyes when the player collides with them.
Added tags that correspond to a loose objects texture which allows me to adjust the sound effect for collision.